+++ lib/app/presentation/pages/group/groups_page.dart
@@ -9,7 +9,7 @@ import '../../../application/usecase/group/group_usecase.dart';
 import '../../../application/usecase/group/state/view_groups_provider.dart';
 import '../../../domain/group/entity/group_tile.dart';
 import '../../components/importer.dart';
-import '../../routes/src/routes_data.dart';
+import '../../routes/importer.dart';
 import '../error/components/error_view.dart';
 import '../item/components/list_loader_view.dart';
 import '../presentation_mixin.dart';
